refactor: mock injector
