+++ packages/addons/__tests__/node/inde.test.ts
@@ -1,4 +1,4 @@
-import { createNodeInjector } from '../../../../tools/dev-tool/src/injector-helper';
+import { createNodeInjector } from '@opensumi/ide-dev-tool/src/mock-injector';
 import { MockInjector } from '../../../../tools/dev-tool/src/mock-injector';
 import { AddonsModule } from '../../src/node';
