+++ packages/collaboration/__tests__/node/service.test.ts
@@ -3,7 +3,7 @@
 import * as Y from 'yjs';
 
 import { INodeLogger } from '@opensumi/ide-core-node';
-import { createNodeInjector } from '@opensumi/ide-dev-tool/src/injector-helper';
+import { createNodeInjector } from '@opensumi/ide-dev-tool/src/mock-injector';
 import { MockInjector } from '@opensumi/ide-dev-tool/src/mock-injector';
 import { IFileService } from '@opensumi/ide-file-service';
 import { FileService } from '@opensumi/ide-file-service/src/node';
