+++ packages/core-node/__tests__/common-module/index.test.ts
@@ -2,7 +2,7 @@ import { ICommonServer } from '@opensumi/ide-core-common';
 import { ServerCommonModule } from '@opensumi/ide-core-node';
 import { CommonServer } from '@opensumi/ide-core-node/lib/common-module/common.server';
 
-import { createNodeInjector } from '../../../../tools/dev-tool/src/injector-helper';
+import { createNodeInjector } from '@opensumi/ide-dev-tool/src/mock-injector';
 import { MockInjector } from '../../../../tools/dev-tool/src/mock-injector';
 
 describe('NodeLogger', () => {
