+++ packages/file-search/__tests__/node/file-search.service.test.ts
@@ -2,7 +2,7 @@ import path from 'path';
 
 import { CancellationTokenSource } from '@opensumi/ide-core-common';
 import { FileUri, AppConfig, INodeLogger, NodeLogger } from '@opensumi/ide-core-node';
-import { createNodeInjector } from '@opensumi/ide-dev-tool/src/injector-helper';
+import { createNodeInjector } from '@opensumi/ide-dev-tool/src/mock-injector';
 import { ProcessModule } from '@opensumi/ide-process/lib/node';
 
 import { IFileSearchService } from '../../src';
