+++ packages/file-service/__tests__/node/index.test.ts
@@ -2,16 +2,16 @@ import os from 'os';
 import path from 'path';
 
 import * as fse from 'fs-extra';
-import { JSDOM } from 'jsdom';
-const { window } = new JSDOM('');
-(global as any).UIEvent = window.UIEvent;
+// import { JSDOM } from 'jsdom';
+// const { window } = new JSDOM('');
+// (global as any).UIEvent = window.UIEvent;
 
 import { isWindows } from '@opensumi/ide-core-common';
 import { FileUri, URI } from '@opensumi/ide-core-node';
 import { expectThrowsAsync } from '@opensumi/ide-core-node/__tests__/helper';
 import { MockInjector } from '@opensumi/ide-dev-tool/src/mock-injector';
 
-import { createNodeInjector } from '../../../../tools/dev-tool/src/injector-helper';
+import { createNodeInjector } from '@opensumi/ide-dev-tool/src/mock-injector';
 import { FileSystemWatcherServer } from '../../lib/node/recursive/file-service-watcher';
 import { FileChangeType, IFileService } from '../../src/common';
 import { FileService, FileServiceModule } from '../../src/node';
