+++ packages/logs-core/__tests__/node/log.service.test.ts
@@ -5,7 +5,7 @@ import temp from 'temp';
 
 import { toLocalISOString, ILogService } from '@opensumi/ide-core-common';
 import { AppConfig } from '@opensumi/ide-core-node';
-import { createNodeInjector } from '@opensumi/ide-dev-tool/src/injector-helper';
+import { createNodeInjector } from '@opensumi/ide-dev-tool/src/mock-injector';
 
 import { LogLevel, SupportLogNamespace, ILogServiceManager } from '../../src/common';
 import { LogServiceModule } from '../../src/node';
