+++ packages/remote-opener/__tests__/node/opener.client.test.ts
@@ -1,6 +1,6 @@
 import { INodeLogger } from '@opensumi/ide-core-node';
 
-import { createNodeInjector } from '../../../../tools/dev-tool/src/injector-helper';
+import { createNodeInjector } from '@opensumi/ide-dev-tool/src/mock-injector';
 import { IRemoteOpenerClient, IRemoteOpenerService, RemoteOpenerClientToken } from '../../src/common';
 import { RemoteOpenerClientImpl } from '../../src/node/opener.client';
