+++ tools/dev-tool/__tests__/mock-injector.test.ts
@@ -1,6 +1,7 @@
 import { Injectable } from '@opensumi/di';
 
-import { createBrowserInjector, createNodeInjector } from '../src/injector-helper';
+import { createBrowserInjector } from '../src/injector-helper';
+import { createNodeInjector } from '../src/mock-injector';
 import { MockInjector } from '../src/mock-injector';
 
 describe('mock-injector test', () => {
