+++ tools/dev-tool/src/injector-helper.ts
@@ -14,7 +14,6 @@ import {
   OS,
   getDebugLogger,
 } from '@opensumi/ide-core-common';
-import { INodeLogger, NodeModule, ServerApp } from '@opensumi/ide-core-node';
 
 import { MockLogger, MockLoggerManageClient, MockLoggerService } from '../../../packages/core-browser/__mocks__/logger';
 import { useMockStorage } from '../../../packages/core-browser/__mocks__/storage';
@@ -133,35 +132,5 @@ export function createBrowserInjector(modules: Array<ConstructorOf<BrowserModule
   afterAll(() => {
     app.injector.disposeAll();
   });
-  return app.injector as MockInjector;
-}
-
-function getNodeMockInjector() {
-  const injector = new MockInjector();
-  injector.addProviders(
-    {
-      token: ILoggerManagerClient,
-      useClass: MockLoggerManageClient,
-    },
-    {
-      token: ILogServiceManager,
-      useClass: MockLoggerService,
-    },
-    {
-      token: INodeLogger,
-      useValue: getDebugLogger(),
-    },
-  );
-  return injector;
-}
-
-export function createNodeInjector(modules: Array<ConstructorOf<NodeModule>>, inj?: Injector): MockInjector {
-  const injector = inj || getNodeMockInjector();
-  const app = new ServerApp({ modules, injector } as any);
-
-  afterAll(() => {
-    app.injector.disposeAll();
-  });
-
   return app.injector as MockInjector;
 }
