+++ tools/dev-tool/src/mock-injector.ts
@@ -1,5 +1,7 @@
 import { ConstructorOf, CreatorStatus, Injector, InstanceOpts, Token, TokenResult } from '@opensumi/di';
-import { CommandRegistry } from '@opensumi/ide-core-common';
+import { MockLoggerManageClient, MockLoggerService } from '@opensumi/ide-core-browser/__mocks__/logger';
+import { CommandRegistry, ILogServiceManager, ILoggerManagerClient, getDebugLogger } from '@opensumi/ide-core-common';
+import { INodeLogger, NodeModule, ServerApp } from '@opensumi/ide-core-node';
 
 export class MockInjector extends Injector {
   private mockMap = new Map<Token, [any, any][]>();
@@ -94,4 +96,34 @@ export function mockService<T = any>(target: Partial<T>): any {
       return t[p];
     },
   });
+}
+
+export function getNodeMockInjector() {
+  const injector = new MockInjector();
+  injector.addProviders(
+    {
+      token: ILoggerManagerClient,
+      useClass: MockLoggerManageClient,
+    },
+    {
+      token: ILogServiceManager,
+      useClass: MockLoggerService,
+    },
+    {
+      token: INodeLogger,
+      useValue: getDebugLogger(),
+    },
+  );
+  return injector;
+}
+
+export function createNodeInjector(modules: Array<ConstructorOf<NodeModule>>, inj?: Injector): MockInjector {
+  const injector = inj || getNodeMockInjector();
+  const app = new ServerApp({ modules, injector } as any);
+
+  afterAll(() => {
+    app.injector.disposeAll();
+  });
+
+  return app.injector as MockInjector;
 }
